% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{adjusted_mi}
\alias{adjusted_mi}
\title{Calculates adjusted (normalized) Mutual Information}
\usage{
adjusted_mi(prediction, truth)
}
\arguments{
\item{prediction}{the predicted cluster assignments for each cell}

\item{truth}{the true cluster assignments for each cell}
}
\value{
the adjusted Mutual Information for the prediction/truth pair
}
\description{
Calculates the mutual information corrected for a hypergeometric-distributed
background. Code comes from \url{http://stackoverflow.com/a/38010916/7195775}.
}
