% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{assign_status}
\alias{assign_status}
\title{Calculates the truth table for a branch assignment prediction.}
\usage{
assign_status(prediction, truth)
}
\arguments{
\item{prediction}{predicted cluster assignments}

\item{truth}{true cluster assignments}
}
\value{
the truth table for this prediction given the truth: (TP, TN, FN, FP)
}
\description{
Goes over all pairs of points in the prediction and truth and classifies
them:
- TP: pair of points in same cluster in prediction, truth
- TN: pair of points in diff. cluster in prediction, truth
- FN: pair of points in diff. cluster in prediction, same in truth
- FP: pair of points in same cluster in prediction, diff. in truth
NA values in the predictions are considered FNs.
}
