% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_trees.R
\name{branch_order}
\alias{branch_order}
\title{Orders branches in current tree zone}
\usage{
branch_order(x, zone, parents, branch_orientation)
}
\arguments{
\item{x}{the current branch endpoints}

\item{zone}{which timezone we are at (see PROSSTT)}

\item{parents}{the parent of each branch}

\item{branch_orientation}{the offsets}
}
\value{
the branches of the current timezone, ordered correctly
}
\description{
Orders branch endpoints so that adjacent branches have the same parent and 
everyone has different endpoints
}
