% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_method.R
\name{evaluate_method}
\alias{evaluate_method}
\title{Evaluate a lineage tree prediction by all available methods}
\usage{
evaluate_method(method, mbranches, mtimes, cell_params, par_loc,
  returnNA = FALSE)
}
\arguments{
\item{method}{the name of the assessed trajectory inference algorithm}

\item{mbranches}{the predicted branch assignments}

\item{mtimes}{the predicted pseudotime assignments}

\item{cell_params}{a data frame that includes branch and pseudotime labels}

\item{par_loc}{location of the PROSSTT parameter file}

\item{returnNA}{if set to TRUE will return an empty result matrix}
}
\value{
an one-row matrix with the evaluation results for all methods
}
\description{
Run all implemented methods for branch assignment and pseudotime prediction
evaluation.
}
