% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{expected_mi}
\alias{expected_mi}
\title{Calculates the expected mutual information}
\usage{
expected_mi(s1, s2, l1, l2, n)
}
\arguments{
\item{s1}{counts per cluster for the prediction}

\item{s2}{counts per cluster for the truth}

\item{l1}{length of s1}

\item{l2}{length of s2}

\item{n}{number of samples}
}
\value{
expected Mutual Information
}
\description{
Calculates the mutual information background, assuming a hypergeometric 
distribution. Code comes from Stack Overflow 
(\url{http://stackoverflow.com/a/38010916/7195775}).
}
