% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{f_measure}
\alias{f_measure}
\title{Calculates the F measure}
\usage{
f_measure(status, b = 1)
}
\arguments{
\item{status}{a truth table}

\item{b}{determines the balance between precision and recall}
}
\value{
the F measure of the input truth table
}
\description{
Calculates the F measure, which measures the effectiveness of retrieval with
respect to a user who attaches \code{b} times as much importance to recall
\code{R} as precision \code{P}:
}
\details{
\deqn{\frac{(b^2 + 1)\mathrm{P} \mathrm{R}}{b^2 \mathrm{P} + \mathrm{R}}}
}
