% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_trees.R
\name{fan_out}
\alias{fan_out}
\title{Spaces points equally.}
\usage{
fan_out(n)
}
\arguments{
\item{n}{number of points needed}
}
\value{
equispaced replacement candidates
}
\description{
Proposes n equidistantly spaced points.
}
