% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_trees.R
\name{flat_simulation}
\alias{flat_simulation}
\title{Calculate offsets in order to plot a simulation in 2D}
\usage{
flat_simulation(cell_params, param_file, mode = "prosstt")
}
\arguments{
\item{cell_params}{a data frame that includes branch and pseudotime labels}

\item{param_file}{location of the PROSSTT parameter file}

\item{mode}{whether the simulations are produced by PROSSTT directly or by
splatter after the MERLoT script}
}
\value{
the start and end point of each branch in 2D
}
\description{
Orders all cells in a simulation by branch and time label, trying to avoid
overlaps.
}
