% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Creates ggplot2 colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{how many colors to generate}
}
\value{
the n first colors from the ggplot2 palette
}
\description{
Creates colors from the default ggplot2 palette.
}
\examples{
gg_color_hue(3)

}
