% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{goodman_kruskal_index}
\alias{goodman_kruskal_index}
\title{Calculates the Goodman-Kruskal index}
\usage{
goodman_kruskal_index(a, b, weighted = TRUE)
}
\arguments{
\item{a}{ordering of the first sequence}

\item{b}{ordering of the second sequence}

\item{weighted}{calculate the weighted version, which takes distances between
points into account.}
}
\value{
the Goodman-Kruskal of the input truth table
}
\description{
Calculates the Goodman-Kruskal index, a comparison of two ordered sequences,
essentially the fraction of concordant pairs in the two sets if neutral pairs
are ignored.
}
