% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{jaccard}
\alias{jaccard}
\title{Calculates the Jaccard index.}
\usage{
jaccard(status)
}
\arguments{
\item{status}{a truth table}
}
\value{
the Jaccard index of the input truth table
}
\description{
Calculates the Jaccard index, also known as Intersection over Union:
}
\details{
\deqn{\frac{\mathrm{TP}}{\mathrm{TP} + \mathrm{FP} + \mathrm{FN}}}
}
