% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{kendall_index}
\alias{kendall_index}
\title{Calculates the Kendall index}
\usage{
kendall_index(a, b, weighted = TRUE)
}
\arguments{
\item{a}{ordering of the first sequence}

\item{b}{ordering of the second sequence}

\item{weighted}{calculate the weighted version, which takes distances between
points into account.}
}
\value{
the Kendall index of the input truth table
}
\description{
Calculates the Kendall index, a comparison of two ordered sequences,
essentially the fraction of concordant pairs in the two sets. Very similar to
the Goodman-Kruskal index, but it also considers neutral pairs.
}
