% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{map2color}
\alias{map2color}
\title{Translates numbers to colors for plotting.}
\usage{
map2color(x, pal, limits = NULL)
}
\arguments{
\item{x}{the input vector}

\item{pal}{the color palette}

\item{limits}{the limits within x moves. If `NULL`, then defaults to 
`range(x)`}
}
\value{
a vector of the appropriate palette color for every number in x
}
\description{
Bins the input number values in as many bins as there are available colors
and returns a vector of the appropriate color for each input value.
}
