% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{matthews_cor}
\alias{matthews_cor}
\title{Calculates the Matthews Correlation Coefficient.}
\usage{
matthews_cor(status)
}
\arguments{
\item{status}{a truth table}
}
\value{
the Matthews Correlation Coefficient of the input truth table
}
\description{
Calculates the Matthews Correlation Coefficient:
\deqn{\frac{\mathrm{TP}\mathrm{TN} - \mathrm{FP}\mathrm{FN}}{\sqrt{(\mathrm{TP}+\mathrm{FP}) (\mathrm{TP}+\mathrm{FN}) (\mathrm{TN}+\mathrm{FP}) (\mathrm{TN}+\mathrm{FN})}}}
}
