% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_method.R
\name{merlot_trajectories}
\alias{merlot_trajectories}
\title{Finds all paths from the root to an endpoint for MERLoT}
\usage{
merlot_trajectories(tree_name, job, embed = TRUE)
}
\arguments{
\item{tree_name}{suffix of the binary file that contains the predicted tree}

\item{job}{location and prefix of the file that contains the predicted tree}

\item{embed}{flag for MERLoT: embed tree or no}
}
\value{
a list of cell IDs, one for each terminal path
}
\description{
Finds all paths that lead from the lineage tree root to an endpoint and the
cells on them
}
