% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{norm_lib_size}
\alias{norm_lib_size}
\title{Normalizes expression matrix per cell.}
\usage{
norm_lib_size(X)
}
\arguments{
\item{X}{the expression matrix to be normalized}
}
\value{
normalized expression matrix
}
\description{
Normalizes the expression matrix per row (cell) by dividing with the average
library size (row sum) and multiplying with the original row sum.
}
