% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_method.R
\name{on_longest_path}
\alias{on_longest_path}
\title{Calculates truth table for cells in longest overlapping paths}
\usage{
on_longest_path(tree_name, job, tool_function, embed = FALSE)
}
\arguments{
\item{tree_name}{suffix of the binary file that contains the predicted tree}

\item{job}{location and prefix of the file that contains the predicted tree}

\item{tool_function}{function to locate longest trajectory in prediction}

\item{embed}{flag for MERLoT: embed tree or no. Needed here for batch runs}
}
\value{
(flat) truth table
}
\description{
Finds the predicted trajectory with the largest overlap with the simulated
longest path and then calculates a truth table for this prediction.
}
