% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_flat_tree}
\alias{plot_flat_tree}
\title{Plots a simulation in 2D}
\usage{
plot_flat_tree(cell_params, branch_orientation, prediction, pcex = 1,
  plot_title = "", col_pal = NA, time_step = 50)
}
\arguments{
\item{cell_params}{data frame that contains branch and pseudotime labels}

\item{branch_orientation}{the coordinates of each branch}

\item{prediction}{predicted pseudotime or branch labels}

\item{pcex}{controls cell size at plotting}

\item{plot_title}{the title of the plot}

\item{col_pal}{a color palette; will be generated if absent}

\item{time_step}{the step size for the x axis label (pseudotime)}
}
\value{
None
}
\description{
Arranges simulation cells according to branch and pseudotime and 
colors them by their predicted branch or pseudotime.
}
