% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{randInd_manual}
\alias{randInd_manual}
\title{Calculates the (unweighted) Rand index}
\usage{
randInd_manual(status)
}
\arguments{
\item{status}{a truth table}
}
\value{
the Rand index of the input truth table
}
\description{
Calculates the (unweighted) Rand index:
\deqn{\frac{\mathrm{TP} + \mathrm{TN}}{\mathrm{TP} + \mathrm{FP} + \mathrm{TN} + \mathrm{FN}}}
}
