% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_trees.R
\name{read_prosstt_topology}
\alias{read_prosstt_topology}
\title{Read the topology of a PROSSTT-style simulation}
\usage{
read_prosstt_topology(param_file)
}
\arguments{
\item{param_file}{location of the parameter file}
}
\value{
the simulation topology as a matrix
}
\description{
Reads a parameter file and returns the topology as pairs of connected
branches
}
