% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_trees.R
\name{time_ordered_branch}
\alias{time_ordered_branch}
\title{Reorders the cells of a branch according to their pseudotime}
\usage{
time_ordered_branch(b, branches, times)
}
\arguments{
\item{b}{the branch which to order}

\item{branches}{the branch labels of all cells}

\item{times}{the pseudotime labels of all cells}
}
\value{
the cells of branch b ordered by pseudotime
}
\description{
Reorders the cells of a branch according to their pseudotime
}
