% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_trees.R
\name{to_flip}
\alias{to_flip}
\title{Which of the duplicates should be flipped}
\usage{
to_flip(duplics)
}
\arguments{
\item{duplics}{a vector of duplicates}
}
\value{
which duplicates to flip
}
\description{
Given a vector of (possible) duplicates, decide which to flip
}
